chore: migrate logic
