+++ frontend/src/views/sql-editor/SecondarySidebar/SecondaryGutterBar.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="flex flex-col items-stretch justify-start text-sm gap-y-1">
     <div
+      v-if="showInfoPane"
       class="gutter-bar--tab"
       :class="[activeTab === 'INFO' && 'gutter-bar--tab-active']"
       @click="handleClickTab('INFO')"
@@ -25,7 +26,10 @@
 </template>
 
 <script setup lang="ts">
-import { computed } from "vue";
+import { computed, watch } from "vue";
+import { useInstanceV1Store, useTabStore } from "@/store";
+import { UNKNOWN_ID } from "@/types";
+import { instanceV1HasAlterSchema, isDisconnectedTab } from "@/utils";
 import { TabView, useSecondarySidebarContext } from "./context";
 
 const { show, tab } = useSecondarySidebarContext();
@@ -37,6 +41,40 @@ const activeTab = computed(() => {
   return tab.value;
 });
 
+const tabStore = useTabStore();
+
+const isDisconnected = computed(() => {
+  return isDisconnectedTab(tabStore.currentTab);
+});
+
+const isSchemalessInstance = computed(() => {
+  if (isDisconnected.value) {
+    return false;
+  }
+  const { instanceId } = tabStore.currentTab.connection;
+
+  if (instanceId === String(UNKNOWN_ID)) {
+    return false;
+  }
+
+  const instance = useInstanceV1Store().getInstanceByUID(instanceId);
+
+  return !instanceV1HasAlterSchema(instance);
+});
+
+const showInfoPane = computed(() => {
+  if (isDisconnected.value) {
+    return false;
+  }
+
+  const conn = tabStore.currentTab.connection;
+  if (conn.databaseId === String(UNKNOWN_ID)) {
+    return false;
+  }
+
+  return !isSchemalessInstance.value;
+});
+
 const handleClickTab = (target: TabView) => {
   if (target === activeTab.value) {
     show.value = false;
@@ -46,11 +84,30 @@ const handleClickTab = (target: TabView) => {
   tab.value = target;
   show.value = true;
 };
+
+watch(
+  showInfoPane,
+  (show) => {
+    if (!show && tab.value === "INFO") {
+      tab.value = "SHEET";
+    }
+  },
+  { immediate: true }
+);
+
+watch(
+  () => tabStore.currentTab.id,
+  () => {
+    if (showInfoPane.value) {
+      tab.value = "INFO";
+    }
+  }
+);
 </script>
 
 <style lang="postcss" scoped>
 .gutter-bar--tab {
-  @apply writing-vertical-rl px-1 py-4 border-y bg-gray-50 cursor-pointer;
+  @apply writing-vertical-rl px-1 py-4 border-y bg-gray-50 cursor-pointer select-none;
 }
 .gutter-bar--tab:first-child {
   @apply border-t-0;
